Revert "Type-safe R entry parsing (#740)"
