--- src/r-bridge/lang-4.x/ast/parser/main/internal/values/normalize-symbol.ts
@@ -1,53 +1,0 @@
-import type { NormalizerData } from '../../normalizer-data'
-import { guard } from '../../../../../../../util/assert'
-import { parseLog } from '../../../json/parser'
-import { retrieveMetaStructure } from '../../normalize-meta'
-import { startAndEndsWith } from '../../../../../../../util/strings'
-import type { RSymbol } from '../../../../model/nodes/r-symbol'
-import { isSymbol, RType } from '../../../../model/type'
-import type { NamedJsonEntry } from '../../../json/format'
-
-/**
- * Normalize the given entries as R symbols (incorporating namespace information).
- *
- * The special symbols `T` and `F` are parsed as logic values.
- *
- * @param data    - The data used by the parser (see {@link NormalizerData})
- * @param entries - The json entries to extract the meta-information from
- *
- * @returns The parsed symbol (with populated namespace information) or `undefined` if the given object is not a symbol.
- */
-export function tryNormalizeSymbol(data: NormalizerData, entries: readonly NamedJsonEntry[]): RSymbol | undefined {
-	guard(entries.length > 0, 'to parse symbols we need at least one object to work on!')
-	parseLog.debug('trying to parse symbol')
-
-	let location, content, namespace
-
-	if(entries.length === 1 && isSymbol(entries[0].name)) {
-		const meta  = retrieveMetaStructure(entries[0].content)
-		location    = meta.location
-		content     = meta.content
-		namespace   = undefined
-	} else if(entries.length === 3 && isSymbol(entries[2].name)) {
-		const meta  = retrieveMetaStructure(entries[2].content)
-		location    = meta.location
-		content     = meta.content
-		namespace   = retrieveMetaStructure(entries[0].content).content
-	} else {
-		return undefined
-	}
-
-	return {
-		type:    RType.Symbol,
-		namespace,
-		location,
-		// remove backticks from symbol
-		content: startAndEndsWith(content, '`') ? content.substring(1, content.length - 1) : content,
-		lexeme:  content,
-		info:    {
-			fullRange:        data.currentRange,
-			additionalTokens: [],
-			fullLexeme:       data.currentLexeme
-		}
-	}
-}
