--- src/r-bridge/lang-4.x/ast/parser/main/normalizer-data.ts
@@ -1,26 +1,0 @@
-import type { MergeableRecord } from '../../../../../util/objects'
-import type { SourceRange } from '../../../../../util/range'
-
-/**
- * Contains all information populated and present during parsing and normalization of the R AST.
- */
-export interface NormalizerData extends MergeableRecord {
-	/**
-   * The currently active source range during parsing, i.e. the full range of the current element.
-   */
-	currentRange:  SourceRange | undefined
-	/**
-   * The currently active lexeme during parsing, i.e. the full lexeme of the current element.
-   */
-	currentLexeme: string | undefined
-}
-
-/**
- * Thrown if the given input is not valid/contains unexpected elements.
- */
-export class ParseError extends Error {
-	constructor(message: string) {
-		super(message)
-		this.name = 'ParseError'
-	}
-}
